+++ particula/next/gas/vapor_pressure_builders.py
@@ -1,8 +1,8 @@
 """Builders to create vapor pressure models for gas species."""
 
-from abc import ABC, abstractmethod
-from typing import Optional, Any
+from typing import Optional
 import logging
+from particula.next.abc_builder import BuilderABC
 from particula.next.gas.vapor_pressure_strategies import (
     AntoineVaporPressureStrategy,
     ClausiusClapeyronStrategy,
@@ -14,124 +14,7 @@ from particula.util.input_handling import convert_units  # type: ignore
 logger = logging.getLogger("particula")
 
 
-class BuilderBase(ABC):
-    """Abstract base class for builders with common methods to check keys and
-    set parameters from dict."""
-
-    def __init__(
-        self,
-        required_parameters: Optional[list[str]] = None
-    ):
-        self.required_parameters = required_parameters or []
-
-    def check_keys(
-        self,
-        parameters: dict[str, Any],
-    ):
-        """Check if the keys you want to set are present in the
-        self.required_parameters dictionary.
-
-        Args:
-        ----
-        - parameters (dict): The parameters dictionary to check.
-        - required_keys (list): List of required keys to be checked in the
-        parameters.
-
-        Returns:
-        -------
-        - None
-
-        Raises:
-        ------
-        - ValueError: If you are trying to set an invalid parameter.
-        """
-        # check if all required keys are present
-        missing = [p for p in self.required_parameters if p not in parameters]
-        if missing:
-            logger.error(
-                "Missing required parameter(s): %s", ', '.join(missing))
-            raise ValueError(
-                f"Missing required parameter(s): {', '.join(missing)}")
-        # check if all keys in parameters are valid, account for _units
-        valid_keys = set(
-            self.required_parameters
-            + [f"{key}_units" for key in self.required_parameters]
-        )
-        key_to_set = [key for key in parameters
-                      if key not in valid_keys]
-        if key_to_set:
-            logger.error(
-                "Trying to set an invalid parameter(s) '%s'. "
-                "The valid parameter(s) '%s'.",
-                key_to_set, valid_keys
-            )
-            raise ValueError(
-                f"Trying to set an invalid parameter(s) '{key_to_set}'."
-                f" The valid parameter(s) '{valid_keys}'."
-            )
-
-    def set_parameters(self, parameters: dict[str, Any]):
-        """Set coefficients from a dictionary including optional units.
-
-        Args:
-        ----
-        - parameters (dict): The parameters dictionary to set.
-
-        Returns:
-        -------
-        - self: The builder object with the set parameters.
-
-        Raises:
-        ------
-        - ValueError: If any required key is missing.
-        - Warning: If using default units for any parameter.
-        """
-        self.check_keys(parameters)  # check if all required keys are present
-        for key in self.required_parameters:  # set the parameters
-            unit_key = f'{key}_units'
-            if unit_key in parameters:
-                # build the set call set with units, from keys
-                # e.g. self.set_a(params['a'], params['a_units'])
-                getattr(self, f'set_{key}')(
-                    parameters[key], parameters[unit_key]
-                )
-            else:
-                logger.warning(
-                    "Using default units for coefficient '%s'.", key)
-                # build set call, e.g. self.set_a(params['a'])
-                getattr(self, f'set_{key}')(parameters[key])
-        return self
-
-    def pre_build_check(self):
-        """Check if all required attribute parameters are set before building.
-
-        Returns:
-        -------
-        - None
-
-        Raises:
-        ------
-        - ValueError: If any required parameter is missing.
-        """
-        missing = [p for p in self.required_parameters
-                   if getattr(self, p) is None]
-        if missing:
-            logger.error(
-                "Required parameter(s) not set: %s", ', '.join(missing))
-            raise ValueError(
-                f"Required parameter(s) not set: {', '.join(missing)}")
-
-    @abstractmethod
-    def build(self) -> Any:
-        """Build and return the strategy object with the set parameters.
-
-        Returns:
-        -------
-        - strategy: The built strategy object.
-        """
-
-
-class AntoineBuilder(BuilderBase):
+class AntoineBuilder(BuilderABC):
     """Builder class for AntoineVaporPressureStrategy. It allows setting the
     coefficients 'a', 'b', and 'c' separately and then building the strategy
     object.
@@ -192,7 +75,7 @@ class AntoineBuilder(BuilderBase):
             self.a, self.b, self.c)  # type: ignore
 
 
-class ClausiusClapeyronBuilder(BuilderBase):
+class ClausiusClapeyronBuilder(BuilderABC):
     """Builder class for ClausiusClapeyronStrategy. This class facilitates
     setting the latent heat of vaporization, initial temperature, and initial
     pressure with unit handling and then builds the strategy object.
@@ -263,41 +146,9 @@ class ClausiusClapeyronBuilder(BuilderBase):
             pressure_initial_units, 'Pa')
         return self
 
-    # def set_parameters(self, parameters: dict):  # type: ignore
-    #     """Set parameters from a dictionary including optional units."""
-    #     required_keys = [
-    #         'latent_heat',
-    #         'temperature_initial',
-    #         'pressure_initial']
-    #     for key in required_keys:
-    #         if key not in parameters:
-    #             raise ValueError(f"Missing coefficient '{key}'.")
-    #         unit_key = f'{key}_units'
-    #         if unit_key in parameters:
-    #             # units provided
-    #             getattr(self, f'set_{key}')(
-    #                 parameters[key], parameters[unit_key]
-    #             )
-    #         else:
-    #             # no units provided
-    #             logger.warning(
-    #                 "Using default units for coefficient '%s'.", key)
-    #             getattr(self, f'set_{key}')(parameters[key])
-    #     return self
-
     def build(self):
         """Build and return a ClausiusClapeyronStrategy object with the set
         parameters."""
-        # if None in [self.latent_heat, self.temperature_initial,
-        #             self.pressure_initial]:
-        #     missing = [
-        #         p for p in [
-        #             'latent_heat',
-        #             'temperature_initial',
-        #             'pressure_initial'] if getattr(
-        #             self,
-        #             p) is None]
-        #     raise ValueError(f"Missing parameters: {', '.join(missing)}")
         self.pre_build_check()
         return ClausiusClapeyronStrategy(
             self.latent_heat,  # type: ignore
@@ -306,7 +157,7 @@ class ClausiusClapeyronBuilder(BuilderBase):
         )
 
 
-class ConstantBuilder(BuilderBase):
+class ConstantBuilder(BuilderABC):
     """Builder class for ConstantVaporPressureStrategy. This class facilitates
     setting the constant vapor pressure and then building the strategy object.
 
@@ -340,35 +191,14 @@ class ConstantBuilder(BuilderBase):
             vapor_pressure_units, 'Pa')
         return self
 
-    # def set_parameters(self, parameters: dict):  # type: ignore
-    #     """Set parameters from a dictionary including optional units."""
-    #     required_keys = ['vapor_pressure']
-    #     for key in required_keys:
-    #         if key not in parameters:
-    #             raise ValueError(f"Missing coefficient '{key}'.")
-    #         unit_key = f'{key}_units'
-    #         if unit_key in parameters:
-    #             # units provided
-    #             getattr(self, f'set_{key}')(
-    #                 parameters[key], parameters[unit_key]
-    #             )
-    #         else:
-    #             # no units provided
-    #             logger.warning(
-    #                 "Using default units for coefficient '%s'.", key)
-    #             getattr(self, f'set_{key}')(parameters[key])
-    #     return self
-
     def build(self):
         """Build and return a ConstantVaporPressureStrategy object with the set
         parameters."""
-        # if self.vapor_pressure is None:
-        #     raise ValueError("Missing parameter: vapor_pressure")
         self.pre_build_check()
         return ConstantVaporPressureStrategy(self.vapor_pressure)
 
 
-class WaterBuckBuilder(BuilderBase):  # pylint: disable=too-few-public-methods
+class WaterBuckBuilder(BuilderABC):  # pylint: disable=too-few-public-methods
     """Builder class for WaterBuckStrategy. This class facilitates
     the building of the WaterBuckStrategy object. Which as of now has no
     additional parameters to set. But could be extended in the future for
